refactor(not-found): implement localized 404 page and loading component for improved user experience
