+++ app/[lang]/[...not-found]/page.tsx
@@ -1,0 +1,7 @@
+import { notFound } from 'next/navigation';
+
+const CatchAllNotFoundPage = () => {
+  return notFound();
+};
+
+export default CatchAllNotFoundPage;
